requirements.txt
fastapi
uvicorn[standard]
numpy
python-dotenv
